\name{FREQ-package}
\alias{FREQ-package}
\alias{FREQ}
\docType{package}
\title{
FREQ: Estimating population size and related parameters from capture frequencies
}
\description{
Real capture frequencies will be fitted to various distributions which provide the basis for estimating population sizes, their standard error, and symmetric as well as asymmetric confidence intervalls. }
\details{
\tabular{ll}{
Package: \tab FREQ\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-22\cr
License: \tab GPL\cr
}
This package uses input data from capture frequencies. There are no limitations on the number of capture occassions.
}
\author{
Annegret Grimm & Klaus Henle

Maintainer: Annegret Grimm, <annegret.grimm@ufz.de>
}
\references{
Grimm, A. (submitted to PLOS One) Reliability of different mark-recapture methods for population size estimation tested with field data from populations of known size.

Henle, K. (1990) Population ecology and life history of the arboreal gecko Gehyra variegata in arid Australia. Herpetological Monographs, 4, 30-60.

Seber, GAF. (1982) The estimation of animal abundance and related parameters. Second edition. Griffin, London.
}
\keyword{ population size }
\keyword{ capture frequencies }
\keyword{ Poisson distribution }
\keyword{ negative binomial distribution }
\keyword{ (truncated) geometric distribution }
